requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==4.0.1
passlib[bcrypt]
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
ecdsa==0.19.2
fastapi==0.135.3
frozenlist==1.8.0
greenlet==3.4.0
h11==0.16.0
idna==3.11
Mako==1.3.11
MarkupSafe==3.0.3
multidict==6.7.1
passlib==1.7.4
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
requests==2.33.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
yarl==1.23.0
email-validator==2.0.0